requirements.txt
pytelegrambotapi
python-dotenv
aiogram
aioschedule